SimpleButton: Move SimpleButton implementation
