+++ Orange/widgets/utils/overlay.py
@@ -14,11 +14,13 @@ from collections import namedtuple
 
 from AnyQt.QtWidgets import (
     QHBoxLayout, QPushButton, QLabel, QSizePolicy, QStyle, QAbstractButton,
-    QStyleOptionButton, QStylePainter, QFocusFrame, QWidget, QStyleOption
+    QWidget, QStyleOption
 )
 from AnyQt.QtGui import QIcon, QPixmap, QPainter
 from AnyQt.QtCore import Qt, QSize, QRect, QPoint, QEvent
 from AnyQt.QtCore import pyqtSignal as Signal, pyqtSlot as Slot
+
+from .buttons import SimpleButton
 
 
 class OverlayWidget(QWidget):
@@ -196,60 +198,6 @@ class OverlayWidget(QWidget):
         self.__widget = None
         if self.isVisible():
             self.hide()
-
-
-class SimpleButton(QAbstractButton):
-    """
-    A simple icon button widget.
-    """
-    def __init__(self, parent=None, **kwargs):
-        super().__init__(parent, **kwargs)
-        self.__focusframe = None
-
-    def focusInEvent(self, event):
-        # reimplemented
-        event.accept()
-        self.__focusframe = QFocusFrame(self)
-        self.__focusframe.setWidget(self)
-
-    def focusOutEvent(self, event):
-        # reimplemented
-        event.accept()
-        self.__focusframe.deleteLater()
-        self.__focusframe = None
-
-    def sizeHint(self):
-        # reimplemented
-        self.ensurePolished()
-        iconsize = self.iconSize()
-        icon = self.icon()
-        if not icon.isNull():
-            iconsize = icon.actualSize(iconsize)
-        return iconsize
-
-    def minimumSizeHint(self):
-        # reimplemented
-        return self.sizeHint()
-
-    def paintEvent(self, event):
-        # reimplemented
-        painter = QStylePainter(self)
-        option = QStyleOptionButton()
-        option.initFrom(self)
-        option.icon = self.icon()
-        option.iconSize = self.iconSize()
-
-        icon = self.icon()
-
-        if not icon.isNull():
-            if option.state & QStyle.State_Active:
-                mode = (QIcon.Normal if option.state & QStyle.State_MouseOver
-                        else QIcon.Active)
-            else:
-                mode = QIcon.Disabled
-            pixmap = icon.pixmap(option.iconSize, mode, )
-
-            painter.drawItemPixmap(option.rect, Qt.AlignCenter, pixmap)
 
 
 class MessageWidget(QWidget):
